First version with pages and altitude sound.
A lot of code from ColorVario::Controller is moved to the new
ColorVario::UserInterface class.

Since this is the first version with two sound sources, the logic around mute
and unmute needs to be reworked.

The old icon box is replaced with two toolbars with real push buttons. The new
push buttons have nice animated clicks and you have to click them properly.
Previously you could swipe your finger across the screen and if you lifted it
on top of one of the icons you sent a signal.

A lot of stuff needs to be fixed (I have a pile of paper on my desk) but I
wanted to commit this version before messing with the mute logic.