Finally, I think I have nailed the sound muting and interruptions by other sound sources. The feedback through the old muted signal caused a lot of problems and this feedback is removed completely.
The locked/unlocked signal is removed from Display and Cpu, just as the
muted/unmuted signal is removed from ValueAudio. To much feedback can cause
trouble and it is not really needed as long as the operation is controlled by
one button.

The color ring is made a little bit thinner.

A small fix to Variometer that avoids sending one crazy value for the first
sensor reading. Cannot calculate climb before we have two values.

The Bluetooth icons now use the same color that is used by the official
Bluetooth icons.

The quit button is darker and the selected version is lighter than the
unselected version. That is a bit strange in my eyes but the same logic is
used by many apps when using a dark theme. For lighter themes they do the
opposite. Furthermore, Oliver liked the lighter selected button better than
the darker I had made first. ;-)