The coolest thing in this commit is that it includes animations for the first time. Both the new menu and the pages are now shown using animations.
The upper and lower toolbars are replaced with a single menu, shown on top.
At the moment it is shown or hidden just by touching the screen, but the plan
is to look into gestures. The new menu contains three buttons and a status icon
for location.

The stop offsets for the color ring needed a small modification. When stopping
at 1 the entire rectangle was filled, and that did not look good when animating
the page shifting. The trick is to add a transparent stop offset at 0.99.

Colors in the ini file now contains the opacity (the alpha value). This make
sense if we decide to change the background color to something else than black.
Furthermore, the default color values are given using floating point values
from 0 to 1. The code that reads the color value still accepts integer values
from 0 to 255. If one of the four values are above 1.5 the integers are
assumed.

When defining off zones for colors, transparent is used instead of black.
Again, this makes more sense because we can change the background to something
else than black.

Location is no longer shown on a button. Instead it is shown as before using an
ImageView. Since the final icon is so small I have reduced the png size to get
a better looking final icon.