A lot of new functionality in class Color:   - Conversions   - Color palettes   - Color blending   - Alpha blending   - Tint and shade   - Gamma correction
Colors can now be read from several different formats. A description of the
formats is given in ColorVario.ini.

A new spreadsheet (doc/Colors.xlsm) is made for testing the blending algorithms
in class Color. In addition, it contains the default climb, altitude and user
interface colors.

The color interpolation in ValueColor now uses the new blending algorithm in
Color. It will not produce the same results as the old one but with the current
default colors, no one can tell.

I have experimented with TextLabel autoscroll but I think it is only horizontal
scrolling that is supported. Furthermore, it only works when size is not set,
probably because size is the equal to the "natural size". To scroll vertically,
we probably need to use another scrollable actor.

Several new functions are added to TextLabel. One of the best is that it
supports some markup formatting.

The current quit and error page will be replaced by a new dialog box that shows
messages. Therefore, the menu now contains two new buttons; an information
button and a quit button. The plan is that the information button will change
it's icon when an error or warning message exists. This is probably better than
showing the error or warning directly.

I have made the icon canvas bigger to make the final button icons smaller.

Screenshots and icons are moved from source repository to Dropbox.