The Interval class is now lower including and upper excluding to avoid problems when the on/off zones starts or ends exactly at a point value on the color or sound curves.
The interval-finding routines in ValueColor and ValueSound are changed to match
the new Interval logic.

The default colors are slightly changed and now follows the color wheel
exactly. The negative colors are made a little darker by making them 50%
transparent.

The unsigned literals used in Color.h and Color.cpp are not required since the
compiler always converts signed to unsigned when the other operand is unsigned,
hence I have replaced them with ordinary signed literals. This is similar to
using 0 and 1 instead of 0.0 and 1.0 when working with doubles. As long as the
compiler rules are clear it is easier to read the uncluttered code that use the
simplest number form.

A small bug in one of the fromRGBA functions is fixed.