A new ColorVario::Menu class encapsulates the menu buttons and status icons.
The ColorVario::Page class is replaced with separate classes for the different
pages; AltitudePage, ClimbPage and SpeedPage, all subclassed from a new Page
class. ColorVarioPage.h and ColorVarioPage.cpp are still around but will be
deleted, eventually.

My own page strip is replaced with a new PageView that uses ScrollView from
the Dali Toolkit. I started by inheriting ScrollView but needed to connect
a few signals, so I have ended up with inheriting from ConnectionTracker and
ScrollView is just a member. However, I have added a cast operator that casts
PageView to a ScrollView when needed.

New gestures for showing and hiding the menu. Works a lot better than the old
method that just used touch up.

I am beginning to like the new auto keyword and is using it more and more. :-)

A new Stopwatch class is added. I had an idea but ended up not using it. If
we end up not using the Stopwatch class, remove it in a later commit.