Finally, MessageDialog and TextView seems to work ok. MessageDialog is tested with 3 messages.
PushButtons and signals are added to Dialog. The buttons are created but not
added to the stage. Derived classes must customize them and add them to the
stage if they need them. For some reason I was unable to connect signals in
MessageDialog to functions in Dialog. Strange but it might be a problem with
the Dali signals. Anyway, that's the main reason why I added all three buttons
to Dialog.

It's very nice to hide and show dialogs using an animation. In my mind the best
animation collapses them when they are closed and expands them when they are
shown. However, since all the elements are defined as rectangles, minimizing
shows their rectangular nature, and that's not so nice. That's why I am
currently testing the opposite animation. Discuss with Kyrre!

Code for setting an elliptical/circular button background is moved from Menu to
PushButton.

The old and large quit icons are removed. Several new icons are added.