Finally, I decided to change the logic for value labels and value rings. Instead of beeing labels and controls they now contain labels and controls, and their names are changed to value writers and value painters.
The Dialog class now creates all three buttons (accept, reject and close)
fully, but hides them initially. Derived classes must show the buttons they
want. It also has a new signal that is emitted after the show or hide animation
is finished.

A new altitude offset dialog is added to enable the end user to enter an
altitude offset. The offset value is saved in AltitudeOffset.ini and loaded
again when ColorVario starts. This makes it possible to reset altitude to
landing height when visiting the landing field and then shut off ColorVario
until start. The dialog is not finished!

The Variometer class is rewritten to handle the new altitude offset.

The auto keyword is used for all functions that return a signal.

The Callback class is changed to allow calling execute from a const reference.
This was needed by ValueWriter and seems ok. Signal's emit is another story and
should remain non-const.

Layer RaiseToTop is moved after it is added to it's parent. Otherwise it
doesn't make any sense.

The about message is moved to the controller.