The icons are updated. They are smaller to get less distortion and the menu icons now contains their own circular background. App icon is updated based on the speed page. Location icon not yet updated.
PushButton::setEllipticBackground is no longer in use and can be removed, but is kept for now.

AltitudeOffsetDialog is improved. New arrows. Variable increment/decrement. Long press to reset offset or altitude instead of tap. The variable increment/decrement logic must be improved.

New disabled color is added to the Color class.

The gesture detectors are attached to all visible controls that might overlap the wanted controls. The reason for this is that the gesture signals are emitted even if another control is on top as long as this control doesn't handle the gesture itself.

Since I experienced a crash when ValueAudio tried to restart the audio I have rewritten the restart code. However, it needs to be tested. Since this has happened only once, maybe the entire restart code can be removed.