This commit was supposed to be a release candidate for version 2. Everything looked promising on the watch with Tizen 3. Then Samsung delivered Tizen 4 to our Gear S3s and several problems occured. Therefore, we might upgrade to Tizen 4 also in the development files. This will require Tizen 4 on the end user device, hence forcing all users to upgrade. Anyway, this commit might be the last version that runs on Tizen 3.
The following is included in this commit:

GPS handling is improved. A new LocationModule class is added, one level higher
than LocationManager. The name might change in the future, but for now it just
works! :-) GPS is turned on when required (e.g. when showing the speed) and
turned off when not required.

Having a GPS is no longer a required feature. ColorVario will now work even on
devices without a GPS. The end user will get a warning about the consequences
but the vario part will still work.

The geoid functionality is moved from LocationManager to the new LocationModule
to avoid having to reload the geoid data each time the GPS is restartet. The
LocationManager object is still used through a smart pointer. The reasoning
behind this is that if we are running on a device without a GPS I think trying
to create a location manager will crash the application. At least this happened
when I tried to run similar Bluetooth code on the simulator (that doesn't
support Bluetooth).

Bluetooth is also removed both as a required feature and as a privilege. The
privilege needs to be added again when we add Bluetooth support in a future
version. However, adding Bluetooth as a required feature is not recommended
since the basic vario works well without it. It is better to add a warning like
we did for a missing GPS.

ColorVarioController now has it's own timer for starting or stopping the GPS.
Previously it relied on the variometer giving a signal. However, the variometer
gives a signal 10 times a second, and we don't need to start or stop the GPS
that often. The new timer ticks every 5 seconds which is probably ok.

A new AltitudeOffsetWriter is created and used instead of just using a regular
AltitudeWriter. This allowed me to move some of the increment and decrement
logic to this class and simplify the AltitudeOffsetDialog.

The AltitudeOffsetDialog now handles feets. Increment/decrement goes in steps
of two feets, same rounding used when displaying altitude.

ValueWriter now contains it's own timer that will write the title instead of
the value to it's "paper" when the value is late. Typically, this can happen
to the speed label because the GPS response can be slow for a number of
reasons. ValuePainter also has similar logic and paints with transparency if
the value is late.

A new namespace SystemInfo contains functions for checking system cababilities
and is used to check if a GPS is supported on the device.

A small fix in TextView synchronizes pan and wheel scrolling.

I had to create a new certificate after the watch was upgraded to Tizen 4.
This might be the reason why there is a new signature2.xml file. I don't know!
The certificates and signatures are still a mystery.

The default sink colors are changed from 50% transparent to fully opaque. The
transparent versions looked too dim after I skipped my own alpha blending.