First version written for Tizen 4.
To be able to compile for Tizen 4, I had to switch to the newer version of the
compiler.

I have made a few small fixes but there are still a few things left to fix
before we can release this version. The most important is to avoid crashing
after a call has interrupted the sound. This worked perfectly on Tizen 3, but
not now.

Default font size now depends on the user's font preferences. I am not 100%
sure if this have been the case all along but at least the result was more
dramatic now. I had chosen big fonts and the result was that all my value
fields that didn't set their size explicitly became very big and on the speed
page even to big to fit inside their bounding rectangle. All text items have
now got their own fixed font size.

The back button didn't lower the app. Again something that worked in Tizen 3.
To get back this functionality I have added a visible signal in the menu and
is setting or clearing the key input focus for page view each time the menu is
shown or hidden. This should not be necessary, since Dali should automatically
give back the key input focus to the page view when the menu clears it.
However, this is Dali and Tizen! ;-)

It was almost impossible to click on the mute audio button. The reason was that
my synchronizing code was called immediately. Previously, that never happened
until after the click was handled. The reason I am synchronizing the buttons
with the actual status of the audio generators is that I want the mute button
to show muted if someone else has taken the sound stream focus. To fix this I
have added a startedOrStopped variable inside ValueAudio and checks this before
checking or unchecking the mute button. This seems to work ok, but something is
wrong with the sound logic since ColorVario often crashes when an incoming
call ends.

All push buttons had the text label BOTTOM_BTN. In addition, the ha a
background that was perfect for my close button at the bottom of the screen.
Obviously, there are some nice features behind the scene. However, I have not
been able to figure out how to control these features. Can't find anything in
the documentation. For now, I remove this background by setting the background
color to transparent. Luckily, I had made my own PushButton class and could
add these changes to the PushButton::New function, and all my buttons appeared
as before.

I have increased the size of the ColorVario label inside the menu from 8 to 9,
and decreased the size of the location symbol. Less chance that anyone thinks
it's clickable.

I have decreased the font size in the message dialog (TextView). Not sure about
this. Discuss with Kyrre and Einar and consider going back to 8 points.