PrivilegeManager is added to enable the end user to allow access to the GPS and media storage directly in ColorVario. Up to now (for Tizen 4) the access had to be given in the Tizen settings (under Apps | Permissions).
The period for the value late timers in ValueWriter and ValuePainter is
increased to 5 times the sampling interval. In Brasil, Kyrre noticed that the
text "Speed" appeared and disappeared several times, probably due to late
values from the GPS.

A quit timer is added to delay the actual quit by a short period to allow the
user interface to update. This small fix avoids the problem where the quit
button disappeared after it was pressed.