I have tried to find the reason why ColorVario sometimes crash when interrupted by an incoming call. I have not been able to fix this 100% but it seems to be a lot more stable with the small changes I have done.
However, I have found that ColorVario always crashes if it tries to acquire
the sound stream focus during a call, even if the call is done by the connected
phone. This should be fixed by detecting the call and avoid acquiring the
sound stream focus while the call is active.

Instead of synchronizing the mute audio button in UserInterface::setAltitude
I have now created a separate timer and callback for this. It is called once a
second instead of 10 times a second that is standard for setAltitude.

I have removed the code that restarted the audio in ValueAudio::setValue. It
might have happen a few times but it was very rare.

I have removed the code regarding context lost and regained. The reason it was
introduced in the first place was that sometimes the user interface did not
show up after the screen had been off. One idea was that this was beacuse the
context was lost. However, these functions have never been called since nothing
has ended up in the log file. Furthermore, the problem seems to have been
fized in Tizen 4. It has never happened after the watch was upgraded.

I have removed some code used for testing that was kept as comments.