On Tizen 3, ColorVario crashed immediately if we did'nt use the sound stream to acquire focus and stopped audio if someone else acquired the sound stream focus. On Tizen 4, we got into trouble when acquiring sound stream focus while a phone call was active. However, for some mysterious reason, everything works perfectly without acquiring sound stream focus. If we just starts and stops the audio as we see fit, we are able to produce vario sounds even when someone else outputs audio. This works during a ring tone, an alarm and a call, and when the media player plays music. If this is accepted by Samsung it is the best solution for a vario, because you won't loose the vario sound if someone calls you or an alarm is triggered.
In this commit, I have removed all the messy code using the sound stream
acquire and release logic. I have still set the sound stream on the audio
output. I don't know if this is required but I think it is the only way to
say that we will produce media output. Anyway, it seems to work. The
SoundStream class is kept as is, but most of it is unused.